chore: formalize notion of block structure in lib
