+++ packages/lib/src/utils/hash-structure.ts
@@ -1,0 +1,6 @@
+import objectHash from 'object-hash';
+import { BlockStructure } from '../models/block-structure';
+
+export const hashStructure = (structure: BlockStructure) => {
+  return objectHash(structure);
+};
